Lazy-load sections (#13061)
* Lazy-load sections

Previously we eagerly loaded sections for better UX, because the built-in lazy loading of turbo frames only loads them when they become viesable, which meant a waiting time for the visitor for frame to load. Now we load them a bit before so when user scrolls to the frame it's already loaded.

* Update prefetch_lazy_controller.js

* Update prefetch_lazy_controller.js